email_processing/requirements.txt
msal
python-dotenv
httpx

numpy
spacy

fastapi
"uvicorn[standard]"